+++ css/portal-dashboard/all-responses-popup/spotlight-message-dialog.less
