+++ css/portal-dashboard/all-responses-popup/spotlight-student-list-dialog.less
