+++ js/components/portal-dashboard/all-responses-popup/spotlight-student-list-dialog.tsx
@@ -9,7 +9,7 @@ import SpotlightIcon from "../../../../img/svg-icons/spotlight-icon.svg";
 import AssignmentIcon from "../../../../img/svg-icons/assignment-icon.svg";
 import SmallCloseIcon from "../../../../img/svg-icons/small-close-icon.svg";
 
-import css from "../../../../css/portal-dashboard/all-responses-popup/selected-student-list.less";
+import css from "../../../../css/portal-dashboard/all-responses-popup/spotlight-student-list-dialog.less";
 
 interface IProps {
   anonymous: boolean;
